move address encode&decode to util.js
